+++ frontend/src/api/Cat/index.ts
@@ -1,10 +1,4 @@
-import axiosInstance from './index';
-
-export const getUserInfo = async () => {
-  const res = await axiosInstance.get('/users');
-  return res.data;
-};
-
+import axiosInstance from '../index';
 export const getCatInfo = async (name: string) => {
   const res = await axiosInstance.get(`/cat/${name}`);
   return res.data;
@@ -18,25 +12,6 @@ export const deleteCat = async (name: string) => {
 export const updateCat = async (name: string, body: any) => {
   const res = await axiosInstance.patch(`/cat/${name}`, body);
   return res.data;
-};
-
-export const createDiary = async (name: string, data: any) => {
-  const res = await axiosInstance.post(`/diary/${name}`, data);
-  return res.data;
-};
-
-export const getDiary = async (name: string, id: string) => {
-  const res = await axiosInstance.get(`diary/${name}/${id}`);
-  return res.data;
-};
-
-export const deleteDiary = async (name: string, id: string) => {
-  const res = await axiosInstance.delete(`diary/${name}/${id}`);
-  return res.data;
-};
-
-export const logout = async () => {
-  return await axiosInstance.get('/auth/logout');
 };
 
 export const enrollCat = async (data: any) => {
