+++ frontend/src/api/User/index.ts
@@ -1,0 +1,8 @@
+import axiosInstance from '../index';
+export const getUserInfo = async () => {
+  const res = await axiosInstance.get('/users');
+  return res.data;
+};
+export const logout = async () => {
+  return await axiosInstance.get('/auth/logout');
+};
